requirements.txt
google-genai
pydub
gdown
yt-dlp
black
pandas
tqdm
python-dotenv